+++ .pnp.cjs
@@ -207,10 +207,6 @@ const RAW_RUNTIME_STATE =
       "reference": "workspace:packages/sources/1forge"\
     },\
     {\
-      "name": "@chainlink/accuweather-adapter",\
-      "reference": "workspace:packages/sources/accuweather"\
-    },\
-    {\
       "name": "@chainlink/ada-balance-adapter",\
       "reference": "workspace:packages/sources/ada-balance"\
     },\
@@ -743,6 +739,10 @@ const RAW_RUNTIME_STATE =
       "reference": "workspace:packages/sources/nikkei"\
     },\
     {\
+      "name": "@chainlink/nomia-adapter",\
+      "reference": "workspace:packages/sources/nomia"\
+    },\
+    {\
       "name": "@chainlink/nyfed-adapter",\
       "reference": "workspace:packages/sources/nyfed"\
     },\
@@ -817,6 +817,10 @@ const RAW_RUNTIME_STATE =
     {\
       "name": "@chainlink/sochain-adapter",\
       "reference": "workspace:packages/sources/sochain"\
+    },\
+    {\
+      "name": "@chainlink/solactive-adapter",\
+      "reference": "workspace:packages/sources/solactive"\
     },\
     {\
       "name": "@chainlink/solana-view-function-adapter",\
@@ -991,7 +995,6 @@ const RAW_RUNTIME_STATE =
   "ignorePatternData": "(^(?:\\\\.yarn\\\\/sdks(?:\\\\/(?!\\\\.{1,2}(?:\\\\/|$))(?:(?:(?!(?:^|\\\\/)\\\\.{1,2}(?:\\\\/|$)).)*?)|$))$)",\
   "fallbackExclusionList": [\
     ["@chainlink/1forge-adapter", ["workspace:packages/sources/1forge"]],\
-    ["@chainlink/accuweather-adapter", ["workspace:packages/sources/accuweather"]],\
     ["@chainlink/ada-balance-adapter", ["workspace:packages/sources/ada-balance"]],\
     ["@chainlink/aleno-adapter", ["workspace:packages/sources/aleno"]],\
     ["@chainlink/alphachain-adapter", ["workspace:packages/sources/alphachain"]],\
@@ -1160,6 +1163,7 @@ const RAW_RUNTIME_STATE =
     ["@chainlink/nft-blue-chip-adapter", ["workspace:packages/sources/nft-blue-chip"]],\
     ["@chainlink/nftx-adapter", ["workspace:packages/composites/nftx"]],\
     ["@chainlink/nikkei-adapter", ["workspace:packages/sources/nikkei"]],\
+    ["@chainlink/nomia-adapter", ["workspace:packages/sources/nomia"]],\
     ["@chainlink/nyfed-adapter", ["workspace:packages/sources/nyfed"]],\
     ["@chainlink/oanda-adapter", ["workspace:packages/sources/oanda"]],\
     ["@chainlink/observation", ["workspace:packages/observation"]],\
@@ -1189,6 +1193,7 @@ const RAW_RUNTIME_STATE =
     ["@chainlink/set-token-index-adapter", ["workspace:packages/composites/set-token-index"]],\
     ["@chainlink/snowflake-adapter", ["workspace:packages/sources/snowflake"]],\
     ["@chainlink/sochain-adapter", ["workspace:packages/sources/sochain"]],\
+    ["@chainlink/solactive-adapter", ["workspace:packages/sources/solactive"]],\
     ["@chainlink/solana-view-function-adapter", ["workspace:packages/sources/solana-view-function"]],\
     ["@chainlink/spectral-macro-score-adapter", ["workspace:packages/sources/spectral-macro-score"]],\
     ["@chainlink/sportsdataio-adapter", ["workspace:packages/sources/sportsdataio"]],\
@@ -5137,25 +5142,6 @@ const RAW_RUNTIME_STATE =
         "linkType": "SOFT"\
       }]\
     ]],\
-    ["@chainlink/accuweather-adapter", [\
-      ["workspace:packages/sources/accuweather", {\
-        "packageLocation": "./packages/sources/accuweather/",\
-        "packageDependencies": [\
-          ["@chainlink/accuweather-adapter", "workspace:packages/sources/accuweather"],\
-          ["@chainlink/ea-bootstrap", "workspace:packages/core/bootstrap"],\
-          ["@chainlink/ea-test-helpers", "workspace:packages/core/test-helpers"],\
-          ["@types/jest", "npm:29.5.14"],\
-          ["@types/node", "npm:22.14.1"],\
-          ["@types/supertest", "npm:2.0.16"],\
-          ["ethers", "npm:5.8.0"],\
-          ["nock", "npm:13.5.6"],\
-          ["supertest", "npm:6.2.4"],\
-          ["tslib", "npm:2.8.1"],\
-          ["typescript", "patch:typescript@npm%3A5.8.3#optional!builtin<compat/typescript>::version=5.8.3&hash=5786d5"]\
-        ],\
-        "linkType": "SOFT"\
-      }]\
-    ]],\
     ["@chainlink/ada-balance-adapter", [\
       ["workspace:packages/sources/ada-balance", {\
         "packageLocation": "./packages/sources/ada-balance/",\
@@ -8245,6 +8231,21 @@ const RAW_RUNTIME_STATE =
         "linkType": "SOFT"\
       }]\
     ]],\
+    ["@chainlink/nomia-adapter", [\
+      ["workspace:packages/sources/nomia", {\
+        "packageLocation": "./packages/sources/nomia/",\
+        "packageDependencies": [\
+          ["@chainlink/nomia-adapter", "workspace:packages/sources/nomia"],\
+          ["@chainlink/external-adapter-framework", "npm:2.7.0"],\
+          ["@types/jest", "npm:29.5.14"],\
+          ["@types/node", "npm:22.14.1"],\
+          ["nock", "npm:13.5.6"],\
+          ["tslib", "npm:2.4.1"],\
+          ["typescript", "patch:typescript@npm%3A5.8.3#optional!builtin<compat/typescript>::version=5.8.3&hash=5786d5"]\
+        ],\
+        "linkType": "SOFT"\
+      }]\
+    ]],\
     ["@chainlink/nyfed-adapter", [\
       ["workspace:packages/sources/nyfed", {\
         "packageLocation": "./packages/sources/nyfed/",\
@@ -8479,6 +8480,7 @@ const RAW_RUNTIME_STATE =
           ["@chainlink/external-adapter-framework", "npm:2.7.0"],\
           ["@types/jest", "npm:29.5.14"],\
           ["@types/node", "npm:22.14.1"],\
+          ["axios", "npm:1.9.0"],\
           ["ethers", "npm:5.8.0"],\
           ["nock", "npm:13.5.6"],\
           ["tslib", "npm:2.4.1"],\
@@ -8779,6 +8781,21 @@ const RAW_RUNTIME_STATE =
           ["nock", "npm:13.5.6"],\
           ["supertest", "npm:6.2.4"],\
           ["tslib", "npm:2.8.1"],\
+          ["typescript", "patch:typescript@npm%3A5.8.3#optional!builtin<compat/typescript>::version=5.8.3&hash=5786d5"]\
+        ],\
+        "linkType": "SOFT"\
+      }]\
+    ]],\
+    ["@chainlink/solactive-adapter", [\
+      ["workspace:packages/sources/solactive", {\
+        "packageLocation": "./packages/sources/solactive/",\
+        "packageDependencies": [\
+          ["@chainlink/solactive-adapter", "workspace:packages/sources/solactive"],\
+          ["@chainlink/external-adapter-framework", "npm:2.7.0"],\
+          ["@types/jest", "npm:29.5.14"],\
+          ["@types/node", "npm:22.14.1"],\
+          ["nock", "npm:13.5.6"],\
+          ["tslib", "npm:2.4.1"],\
           ["typescript", "patch:typescript@npm%3A5.8.3#optional!builtin<compat/typescript>::version=5.8.3&hash=5786d5"]\
         ],\
         "linkType": "SOFT"\