+++ packages/sources/solactive/test/unit/nav.test.ts
@@ -1,0 +1,164 @@
+import { EndpointContext } from '@chainlink/external-adapter-framework/adapter'
+import { calculateHttpRequestKey } from '@chainlink/external-adapter-framework/cache'
+import { metrics } from '@chainlink/external-adapter-framework/metrics'
+import { TransportDependencies } from '@chainlink/external-adapter-framework/transports'
+import { LoggerFactoryProvider } from '@chainlink/external-adapter-framework/util'
+import { makeStub } from '@chainlink/external-adapter-framework/util/testing-utils'
+import { inputParameters } from '../../src/endpoint/nav'
+import { HttpTransportTypes, NavTransport } from '../../src/transport/nav'
+
+const originalEnv = { ...process.env }
+
+const restoreEnv = () => {
+  for (const key of Object.keys(process.env)) {
+    if (key in originalEnv) {
+      process.env[key] = originalEnv[key]
+    } else {
+      delete process.env[key]
+    }
+  }
+}
+
+const log = jest.fn()
+const logger = {
+  fatal: log,
+  error: log,
+  warn: log,
+  info: log,
+  debug: log,
+  trace: log,
+}
+
+const loggerFactory = { child: () => logger }
+
+LoggerFactoryProvider.set(loggerFactory)
+metrics.initialize()
+
+describe('NavTransport', () => {
+  const transportName = 'default_single_transport'
+  const endpointName = 'nav'
+
+  const adapterSettings = makeStub('adapterSettings', {
+    API_ENDPOINT: 'https://clients.solactive.com/api/rest/v1/indices',
+    WARMUP_SUBSCRIPTION_TTL: 10_000,
+    CACHE_MAX_AGE: 90_000,
+    MAX_COMMON_KEY_SIZE: 300,
+    DEFAULT_CACHE_KEY: 'default-cache-key',
+  } as unknown as HttpTransportTypes['Settings'])
+
+  const subscriptionSet = makeStub('subscriptionSet', {
+    getAll: jest.fn(),
+  })
+
+  const subscriptionSetFactory = makeStub('subscriptionSetFactory', {
+    buildSet() {
+      return subscriptionSet
+    },
+  })
+
+  const requester = makeStub('requester', {
+    request: jest.fn(),
+  })
+
+  const responseCache = {
+    write: jest.fn(),
+  }
+  const dependencies = makeStub('dependencies', {
+    requester,
+    responseCache,
+    subscriptionSetFactory,
+  } as unknown as TransportDependencies<HttpTransportTypes>)
+
+  let transport: NavTransport
+
+  const requestKeyForParams = (params: typeof inputParameters.validated) => {
+    const requestKey = calculateHttpRequestKey<HttpTransportTypes>({
+      context: {
+        adapterSettings,
+        inputParameters,
+        endpointName,
+      },
+      data: [params],
+      transportName,
+    })
+    return requestKey
+  }
+
+  beforeEach(async () => {
+    restoreEnv()
+    jest.restoreAllMocks()
+    jest.useFakeTimers()
+
+    transport = new NavTransport()
+
+    await transport.initialize(dependencies, adapterSettings, endpointName, transportName)
+  })
+
+  it('should make the request', async () => {
+    const clientId = 'abc123'
+    const isin = 'A0B1C2D3'
+    const clientPassword = 'hunter2'
+    const expectedNav = 123.45
+
+    process.env.PASSWORD_ABC123 = clientPassword
+
+    const params = makeStub('params', {
+      clientId,
+      isin,
+    })
+    subscriptionSet.getAll.mockReturnValue([params])
+
+    const context = makeStub('context', {
+      adapterSettings,
+      endpointName,
+    } as EndpointContext<HttpTransportTypes>)
+
+    const response = makeStub('response', {
+      response: {
+        data: {
+          level: expectedNav,
+          cost: {},
+        },
+      },
+      timestamps: {},
+    })
+
+    requester.request.mockResolvedValue(response)
+
+    await transport.backgroundExecute(context)
+
+    const expectedRequestConfig = {
+      auth: {
+        username: clientId,
+        password: clientPassword,
+      },
+      baseURL: adapterSettings.API_ENDPOINT,
+      params: params,
+      url: `/${clientId}/${isin}/performance`,
+    }
+    const expectedRequestKey = requestKeyForParams(params)
+
+    const expectedResponse = {
+      data: {
+        result: expectedNav,
+      },
+      result: expectedNav,
+      timestamps: {},
+    }
+
+    expect(requester.request).toHaveBeenCalledWith(
+      expectedRequestKey,
+      expectedRequestConfig,
+      undefined,
+    )
+    expect(requester.request).toHaveBeenCalledTimes(1)
+
+    expect(responseCache.write).toHaveBeenCalledWith(transportName, [
+      {
+        params,
+        response: expectedResponse,
+      },
+    ])
+    expect(responseCache.write).toHaveBeenCalledTimes(1)
+  })
+})